tests/wpt/web-platform-tests/offscreen-canvas/the-offscreen-canvas/context.unrecognised.badname.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: context.unrecognised.badname</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>context.unrecognised.badname</h1>
<p class="desc">getContext with unrecognised context name returns null</p>


<script>
var t = async_test("getContext with unrecognised context name returns null");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var offscreenCanvas2 = new OffscreenCanvas(100, 50);
assert_throws(new TypeError(), function() { offscreenCanvas2.getContext('This is not an implemented context in any real browser'); });

t.done();

});
</script>
